Move Stratis members
